+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [2.5] - Unreleased
 
 - Picture name with Unicode character no longer crash the program.
+- Added a dark mode if the OS is in dark mode.
 
 ## [2.4] - 2022-06-01