+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Picture name with Unicode character no longer crash the program.
 - Added a dark mode if the OS is in dark mode.
 - Changed the default brush size.
+- Added a filter based on color.
 
 ## [2.4] - 2022-06-01