+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added a dark mode if the OS is in dark mode.
 - Changed the default brush size.
 - Added a filter based on color.
+- Changed the max value for brush size to 100.
 
 ## [2.4] - 2022-06-01